# server/flaskcontract/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Contractors collaboration tools",
    "version": "1.0.0",
    "title": "Contractors collaboration tool specification",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "tags": [
    {
      "name": "Contractors API",
      "description": "Contractors negotiation tool"
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "Authentication": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Id": {
        "type": "string",
        "properties": {
          "Id": {
            "type": "string"
          }
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "room_name": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "RoomMember": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "room_id": {
            "type": "string"
          },
          "added_by": {
            "type": "string"
          },
          "added_at": {
            "type": "string"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "room_member_id": {
            "type": "string"
          },
          "room_id": {
            "type": "string"
          },
          "added_at": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Contract": {
        "type": "object",
        "properties": {
          "ContractId": {
            "type": "string"
          },
          "Purpose": {
            "type": "string"
          },
          "StartDate": {
            "type": "string"
          },
          "EffectiveDate": {
            "type": "string"
          },
          "ExecutionDate": {
            "type": "string"
          },
          "ExpireDate": {
            "type": "string"
          },
          "ContractProvider": {
            "type": "string"
          },
          "ContractRequester": {
            "type": "string"
          },
          "DataController": {
            "type": "string"
          },
          "ContractType": {
            "type": "string"
          },
          "ContractStatus": {
            "type": "string"
          },
          "Amendment": {
            "type": "string"
          },
          "ConfidentialityObligation": {
            "type": "string"
          },
          "DataProtection": {
            "type": "string"
          },
          "LimitationOnUse": {
            "type": "string"
          },
          "Medium": {
            "type": "string"
          },
          "MethodOfNotice": {
            "type": "string"
          },
          "NoThirdPartyBeneficiaries": {
            "type": "string"
          },
          "PermittedDisclosure": {
            "type": "string"
          },
          "ReceiptOfNotice": {
            "type": "string"
          },
          "Severability": {
            "type": "string"
          },
          "TerminationForInsolvency": {
            "type": "string"
          },
          "TerminationForMaterialBreach": {
            "type": "string"
          },
          "TerminationOnNotice": {
            "type": "string"
          },
          "Waiver": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/contract/api/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Login",
        "requestBody": {
          "description": "Provide login information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Authentication"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Authentication"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/Authentication"
            }
          }
        }
      }
    },
    "/contract/api/register": {
      "post": {
        "tags": [
          "Register"
        ],
        "summary": "Register new user",
        "requestBody": {
          "description": "Provide new user information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Authentication"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Authentication"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/Authentication"
            }
          }
        }
      }
    },
    "/contract/api/add_room": {
      "post": {
        "tags": [
          "Add Room"
        ],
        "summary": "Adding room information",
        "requestBody": {
          "description": "Room information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Room"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/Room"
            }
          }
        }
      }
    },
    "/contract/api/add_message": {
      "post": {
        "tags": [
          "Add Message"
        ],
        "summary": "Adding new message",
        "requestBody": {
          "description": "Message information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      }
    },
    "/contract/api/add_members": {
      "post": {
        "tags": [
          "Add Room Members"
        ],
        "summary": "Adding room members",
        "requestBody": {
          "description": "Room members information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomMember"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/RoomMember"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/RoomMember"
            }
          }
        }
      }
    },
    "/contract/api/get_user": {
      "get": {
        "tags": [
          "Get current user"
        ],
        "summary": "Get the current user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/message/{roomid}": {
      "get": {
        "tags": [
          "Get message by room id"
        ],
        "summary": "Get message based on a room",
        "parameters": [
          {
            "name": "roomid",
            "in": "path",
            "description": "room id to obtain message",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_user_by_id/{id}": {
      "get": {
        "tags": [
          "Get User By Id"
        ],
        "summary": "Get User by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id to obtain user",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/delete_room/{id}": {
      "delete": {
        "tags": [
          "Delete room"
        ],
        "summary": "Delete room by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "room id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_all_users": {
      "get": {
        "tags": [
          "Get all users"
        ],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/create": {
      "post": {
        "tags": [
          "Create"
        ],
        "summary": "New contract",
        "requestBody": {
          "description": "Contract information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Contract"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/Contract"
            }
          }
        }
      }
    },
    "/contract/api/update/{id}": {
      "put": {
        "tags": [
          "Update"
        ],
        "summary": "Update the contract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "contract id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "description": "Contract information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Contract"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/Contract"
            }
          }
        }
      }
    },
    "/contract/api/get_all_contracts": {
      "get": {
        "tags": [
          "Get all contracts"
        ],
        "summary": "Get all contracts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_contract_by_id/{id}": {
      "get": {
        "tags": [
          "Get contract By Id"
        ],
        "summary": "Get contract by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "contract id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_contract_by_provider/{name}": {
      "get": {
        "tags": [
          "Get contract By provider"
        ],
        "summary": "Get contract by provider",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "provider",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_contract_by_requester/{name}": {
      "get": {
        "tags": [
          "Get contract By requester"
        ],
        "summary": "Get contract by requester",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "requester",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/contract_delete_by_id/{id}": {
      "delete": {
        "tags": [
          "Delete contract"
        ],
        "summary": "Delete contract by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "contract id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/logout": {
      "get": {
        "tags": [
          "Logout"
        ],
        "summary": "Logout",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_rooms": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get all rooms",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_room_by_id/{id}": {
      "get": {
        "tags": [
          "Get room By Id"
        ],
        "summary": "Get room by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "room id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_room_members/{roomid}": {
      "get": {
        "tags": [
          "Get room members"
        ],
        "summary": "Get room members by room id",
        "parameters": [
          {
            "name": "roomid",
            "in": "path",
            "description": "room id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/get_room_member_by_id/{id}": {
      "get": {
        "tags": [
          "Get room member by id"
        ],
        "summary": "Get room member by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "room member id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/delete_room_member/{id}": {
      "delete": {
        "tags": [
          "Delete room member"
        ],
        "summary": "Delete room member by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "room member id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/contract/api/message/delete/{roomid}": {
      "delete": {
        "tags": [
          "Delete message"
        ],
        "summary": "Delete messages by room id",
        "parameters": [
          {
            "name": "roomid",
            "in": "path",
            "description": "room id",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    }
  }
}